requirements.txt
bottle==0.12.13
bottle-websocket==0.2.9
certifi==2016.2.28
gevent==1.2.2
gevent-websocket==0.10.1
greenlet==0.4.12
six==1.11.0
websocket-client==0.44.0
